dummy-profile-service/requirements.txt
grpcio
asyncio
protobuf
python-dotenv
psycopg2
opentelemetry-distro
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-grpc